+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
+* names kwarg to CocoDataset.images / videos
+
 * DelayedImage.finalize now accepts a `nodata` argument, which handles invalid
   data pixels correctly under transformations.